chore (ai): rename CoreMessage to ModelMessage (#6105)
## Background

The name `CoreMessage` was an artifact of needed to introduce a separate
message type. Over time it became clear that we have threee distinct
message types:

- messages used on the client that are displayed to the user
(`UIMessage`)
- messages that are sent to our user-facing APIs (currently
`CoreMessage`)
- messages that are sent to the providers (`LanguageModelV2Message`)

They all have distinct DX considerations and are different by design.

The name `CoreMessage` is not clear here.

## Summary

Rename `CoreMessage` to `ModelMessage` to make the distinction
`UIMessage <> ModelMessage` clear (most users are not exposed to
internal provider messages).

## Future Work
`CoreMessages` and related types and functions are deprecated and will
be removed in AI SDK 6